% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Linear Regression Summary}
\usage{
LR(data, x, y)
}
\arguments{
\item{data}{input aimed data that need to fit linear regression model}

\item{x}{the list of names of independent variables}

\item{y}{name of dependent variable}
}
\value{
the statistics of linear regression
}
\description{
Gets the Linear regression model and some summary result of aimed data
}
\examples{
library(datasets)
data(swiss)
data <- swiss
x <- c("Fertility", "Agriculture", "Examination", "Education")
y <- "Infant.Mortality"
LR(data, x, y)

}
