% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeRegression.R
\name{ridgeRegression}
\alias{ridgeRegression}
\title{A fast implementation of ridge regression without intercept term}
\usage{
ridgeRegression(X, Y, lambdas)
}
\arguments{
\item{X}{A n * p design matrix}

\item{Y}{A size n vector}

\item{lambdas}{A size m vector of non-negative values}
}
\value{
A list containing beta for each lambda
}
\description{
A fast implementation of ridge regression without intercept term
}
\examples{
y = rnorm(100)
x = matrix(rnorm(100 * 20), 100, 20)
fit = ridgeRegression(x, y, seq(0, 1, 0.1))
}
