% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Linear Regression Summary}
\usage{
LR(data, x, y)
}
\arguments{
\item{data}{a data.frame that needed to fit linear regression model and summary}

\item{x}{a vector of names of independent variables}

\item{y}{a vector of name of dependent variable}
}
\value{
the summary of linear regression
}
\description{
This function is used to arrange some outputs of summary results of linear regression model. It contains more useful outputs
and reduces some outputs that is seldom used for rookies. It is user friendly and you will understand that in examples and tutorial.
}
\details{
Get the Linear regression model summary results of aimed data
}
\examples{
library(datasets)
data(swiss)
data <- swiss
x <- c("Fertility", "Agriculture", "Examination", "Education")
y <- "Infant.Mortality"
LR(data, x, y)

}
